+++ src/main/java/com/woowacamp/soolsool/core/auth/config/WebConfig.java
@@ -20,7 +20,6 @@ public class WebConfig implements WebMvcConfigurer {
             .addPathPatterns("/**")
             .excludePathPatterns("/auth/login", "/members")
             .excludePathPatterns("/error", "/css/**", "/assets/**", "/js/**", "/*.ico");
-
     }
 
     @Override
