+++ src/main/java/com/woowacamp/soolsool/core/auth/service/AuthService.java
@@ -1,6 +1,5 @@
 package com.woowacamp.soolsool.core.auth.service;
 
-
 import static com.woowacamp.soolsool.core.member.code.MemberErrorCode.MEMBER_NO_INFORMATION;
 import static com.woowacamp.soolsool.core.member.code.MemberErrorCode.MEMBER_NO_MATCH_PASSWORD;
