+++ src/main/java/com/woowacamp/soolsool/core/member/domain/vo/MemberRoleType.java
@@ -7,11 +7,9 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public enum MemberRoleType {
 
-    CUSTOMER("구매자", "CUSTOMER"),
-    VENDOR("판매자", "VENDOR"),
+    CUSTOMER("구매자"),
+    VENDOR("판매자"),
     ;
 
     private final String type;
-    private final String eType;
-    
 }
